api/requirements.txt
altgraph==0.10.2
appdirs==1.4.0
astroid==1.6.5
backports.functools-lru-cache==1.5
bdist-mpkg==0.5.0
bonjour-py==0.3
certifi==2017.7.27.1
chardet==3.0.4
conekta==2.3.4
configparser==3.5.0
cycler==0.10.0
Cython==0.28.3
diff-match-patch==20121119
Django==1.11.4
django-admin-bootstrapped==2.5.7
django-admin-bootstrapped-compatible==3.1.7
django-allauth==0.32.0
django-appconf==1.0.2
django-cors-headers==2.1.0
django-crispy-forms==1.6.0
django-datetime-widget==0.9.3
django-excel==0.0.9
django-extra-fields==0.9
django-filter==1.0.4
django-import-export==0.5.1
django-markdown-deux==1.0.5
django-pagedown==0.1.1
django-rest-auth==0.9.1
django-rest-framework==0.1.0
Django-Select2==5.8.10
django-simple-history==2.1.1
django-summernote==0.8.6
django-tables2==1.2.6
djangorestframework==3.6.3
djangorestframework-jwt==1.9.0
enum34==1.1.6
et-xmlfile==1.0.1
ez-setup==0.9
futures==3.2.0
gunicorn==19.6.0
httplib2==0.10.3
idna==2.5
isort==4.3.4
jdcal==1.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lml==0.0.1
macholib==1.5.1
Markdown==2.6.8
markdown2==2.3.1
matplotlib==2.2.2
mccabe==0.6.1
meld3==1.0.2
modulegraph==0.10.4
mysql-connector-python-cext==8.0.12
nose==1.3.7
numpy==1.14.5
oauthlib==2.0.2
odfpy==1.3.6
olefile==0.44
openpyxl==2.5.4
packaging==16.8
Pillow==4.2.1
py2app==0.7.3
pycrypto==2.6.1
pyexcel==0.5.3
pyexcel-io==0.4.4
pyexcel-webio==0.1.2
Pygments==2.1.3
PyJWT==1.4.2
pylint==1.9.3
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.1.10
pystan==2.17.1.0
python-dateutil==2.7.3
python-http-client==3.1.0
python-mpd2==1.0.0
python-openid==2.2.5
pytz==2017.2
PyYAML==3.12
requests==2.18.3
requests-oauthlib==0.8.0
ruamel.ordereddict==0.4.13
ruamel.yaml==0.15.40
scipy==0.13.0b1
sendgrid==5.4.1
simplejson==3.11.1
singledispatch==3.4.0.3
six==1.10.0
subprocess32==3.5.2
supervisor==3.2.0
tablib==0.11.4
texttable==0.9.1
unicodecsv==0.14.1
urllib3==1.22
virtualenv==15.1.0
weakrefset==1.0.0
wrapt==1.10.11
xattr==0.6.4
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.1.1
